requirements.txt
jinja2
git+https://github.com/jmarianer/lupa.git#egg=lupa
Pillow
click
mypy
types-Pillow
pycodestyle
bbcode
MarkupSafe
